added roles to redux
