+++ src/types.tsx
@@ -12,7 +12,8 @@ export interface UserState {
     name: string,
     userName: string,
     organizationName: string
-    picture: string
+    picture: string,
+    roles: UserRole[]
 }
 
 
@@ -462,7 +463,6 @@ export type ContentProps = {
     setCollapseMobileCategories: (collapseMobileCategories: boolean) => void;
     setScaleDescOpen: Dispatch<SetStateAction<boolean>>;
     setFirstTimeLogin: Dispatch<SetStateAction<boolean>>;
-    roles: UserRole[];
     setShowFab: Dispatch<SetStateAction<boolean>>;
 };
